Functions

Description

Documentation

rmDups :: (Eq a, Ord a) => [a] -> [a]

Remove duplicate items from a list

padRight :: Int -> Char -> String -> String

Pad a string right with a given char until it is of a certain length

mean :: Integral a => [a] -> Float

Take the mean of a list of Integral values

makeNumberHumanReadable :: Int -> String

Comma-separate a number for human reading

distanceFrom5 :: Int -> Int

Take the minimum distance from a multiple of 5 that a number is

avgDistanceFromMultiplesOf5 :: [Int] -> Float

Take the average distance from a multiple of 5 that a list of numbers are

bestCaptainOption :: Option -> Option

Take the most best option for a team captain

bestCaptainOption' :: Option -> [Option]

Generate a list of options for the team captain, and order them from "worst" to "best"

iterativeApplicationFn :: Eq a => (a -> a) -> [a] -> [[a]]

Apply a function to each item in a list, returning a list of lists containing | the result of each application

iterativeApplicationFn' :: Eq a => (a -> a) -> [a] -> [a] -> [[a]]

Helper function for iterativeApplicationFn

orderOptions :: Option -> Option -> Ordering

Returns whether tps2 should be higher up the list than tps1

orderOptions' :: [Int] -> [Int] -> (Ordering, String)

Helper function for orderOptions

playerTeamToOption :: [(Player, Team)] -> Option

Convert a list of players and teams to an Option

ppOption :: Option -> String

Nicely print an Option

ppOptions :: [Option] -> String

Nicely print a list of Options

putPPOptions :: [Option] -> IO ()

Print a nicely formatted list of Options

popularitySort :: Lineup -> Lineup

Sort a lineup by popularity (frequency) of teams

numOptionsFn :: Lineup -> Int

How many Options will a given Lineup generate

allTeamsFn :: Lineup -> [Team]

Get a list of all Teams in a lineup

popFilter :: Lineup -> Lineup

Filter a Lineup such that it contains only teams that could be a viable Option

lineupToPlayerTeams :: Lineup -> [[(Player, Team)]]

Convert a Lineup to a list of lists of Players and Teams

foldFunction :: [Option] -> [Option]

The function to fold a list of options down into the best one | It generates a list such that it can be printed and give some idea | Of progress

ppSquad :: Lineup -> IO ()

Nicely print a Lineup