Type

Description

Documentation

type Team = String

Team is shorthand for a String - it is just the name of a team

type Position = String

Position is shorthand for a String - it is just the name of a football position

type Player = String

Player is shorthand for a String - it is just the name of a football player

type PlayerTeams = (Player, [Team])

A player and all of their teams

type Lineup = [PlayerTeams]

A full lineup

type LiterateLineup = [(Position, Lineup)]

A full lineup with position designations (only used to instantiate)

type TeamPlayer = (Team, [Player])

A team and a list of all players with that team's chemistry

type Option = [TeamPlayer]

An option for the whole squad's chemistries